+++ rest_framework/mixins.py
@@ -3,9 +3,6 @@ Basic building blocks for generic class based views.
 
 We don't bind behaviour to http method handlers yet,
 which allows mixin classes to be composed in interesting ways.
-
-Eg. Use mixins to build a Resource class, and have a Router class
-    perform the binding of http methods to actions for us.
 """
 from django.http import Http404
 from rest_framework import status
@@ -78,15 +75,17 @@ class UpdateModelMixin(object):
     def update(self, request, *args, **kwargs):
         try:
             self.object = self.get_object()
+            success_status = status.HTTP_200_OK
         except Http404:
             self.object = None
+            success_status = status.HTTP_201_CREATED
 
         serializer = self.get_serializer(data=request.DATA, instance=self.object)
 
         if serializer.is_valid():
             self.pre_save(serializer.object)
             self.object = serializer.save()
-            return Response(serializer.data)
+            return Response(serializer.data, status=success_status)
 
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)